# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Basic Bank System
  description: |-
    aplikasi ini untuk merepresentasikan pengguna, rekening bank, dan transaksi. Pengguna dapat memiliki profil terkait. Setiap rekening bank memiliki saldo dan terkait dengan transaksi sebagai sumber atau tujuan. Ini memungkinkan sistem untuk melakukan pendaftaran pengguna, transfer dana, dan melacak riwayat transaksi.
  contact:
    email: alex@mail.com
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
...
paths:
  /users:
    post:
      tags:
        - Users
      operationId: post-create-new-user
      summary: Create a new user and a new profile
      description: Creates a new user and a new profile with the provided information.
      requestBody:
        $ref: "#/components/requestBodies/CreateUser"
      responses:
        "201":
          description: Successful user creation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          newUser:
                            $ref: "#/components/schemas/UserModel"
                          newProfile:
                            $ref: "#/components/schemas/ProfileModel"
        "400":
          description: Email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "failed"
    get:
      tags:
        - Users
      operationId: get-fetch-all-user
      summary: Get all user
      description: Retrieves a list of all registered user.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Page number for pagination.
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Number of items per page.
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          pagination:
                            $ref: "#/components/schemas/Pagination"
                          users:
                            type: array
                            items:
                              $ref: "#/components/schemas/UserModel"
  /users/{userId}:
    get:
      tags:
        - Users
      operationId: get-fetch-user-detail
      summary: Get user details
      description: Retrieves details of a specific user.
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: The ID of the user to retrieve.
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BasicResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserModel"
        "400":
          description: No user found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "failed"
...
components:
  schemas:
    BasicResponse:
      title: BasicResponse
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: "success"
    UserModel:
      title: User
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "arfin"
        email:
          type: string
          example: "arfin@mail.com"
        password:
          type: string
          example: "$2b$10$wpqr1W8aUNCN7BZZ3o6sz.meX9rB/aItp5Mw/0q7GaRmJbsU2GGbK"
    ProfileModel:
      title: Profile
      type: object
      properties:
        id:
          type: integer
          example: 1
        identity_type:
          type: string
          example: "KTP"
        identity_number:
          type: string
          example: "1234567890"
        address:
          type: string
          example: "Jalan Sudirman No. 123"
        userId:
          type: integer
          example: 1
...
    Pagination:
      title: Pagination
      type: object
      properties:
        links:
          type: object
          properties:
            next:
              type: string
              example: "https://domain.com/users?page=3&limit=10"
            prev:
              type: string
              example: "https://domain.com/users?page=1&limit=10"
        total_items:
          type: integer
          example: 100
  requestBodies:
    CreateUser:
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                example: "arfin"
              email:
                type: string
                example: "arfin@mail.com"
              password:
                type: string
                example: "123456"
              identity_type:
                type: string
                example: "KTP"
              identity_number:
                type: string
                example: "1234567890"
              address:
                type: string
                example: "Jalan Sudirman No. 123"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT